requirements.txt
aiofiles>=0.6.0
aiohttp[speedups]>=3.7.3
beautifulsoup4>=4.9.1
covid>=2.4.0
cowpy>=1.1.0
dnspython>=2.0.0
gitpython>=3.1.11
gdshortener
google-api-python-client>=1.12.8
google-auth-httplib2>=0.0.4
google-auth-oauthlib>=0.4.2
googletrans==3.1.0a0
hachoir>=3.1.1
heroku3>=4.2.3
html-telegraph-poster>=0.2.31
imutils
motor>=2.3.0
numpy
oauth2client>=4.1.3
opencv-python
Pillow==8.2.0
psutil>=5.7.3
pybase64>=1.1.1
git+https://github.com/code-rgb/pyrogram.git@x21
pySmartDL>=1.3.4
python-dotenv>=0.15.0
pytz>=2020.4
PyPDF2
rarfile>=4.0
removebg>=0.4
requests>=2.25.0
git+https://github.com/fnixdev/search-engine-parser
selenium==3.141.0
setuptools>=40.3.0
spamwatch>=0.3.0
speedtest-cli>=2.1.3
stagger>=1.0.0
scikit-image
telegraph>=1.4.1
git+https://github.com/fnixdev/sqlalchemy
pickledb
tgcrypto
urbandict>=0.5
wget>=3.2
wikipedia>=1.4.0
git+https://github.com/ytdl-org/youtube-dl
git+https://github.com/Nekos-life/Async-nekos.life-wrapper
ruamel.yaml==0.17.2
ujson>=4.0.1
youtube-search-python==1.4.2
git+https://github.com/fnixdev/google-images-download.git
feedparser>=6.0.2
